Checking new format of reporting and adding tests.
Adding tests as separate files and launching them as separate processes provides more memory isolation, as opposed to the previous approach of having similar tests in the same process
