+++ tests/output.mjs
@@ -1,4 +1,4 @@
 export function output(items, type, speed) {
 	const itemCount = items / 1000000 + ' million';
-	console.log(`- [${itemCount} items] - {${type}} - ${speed}`);
+	console.log(`| [${itemCount} items] | {${type}} | ${speed} |`);
 }
